Move predicates file to languages folder
