+++ extensions/ql-vscode/src/model-editor/languages/index.ts
@@ -1,0 +1,1 @@
+export * from "./predicates";
